fix: changed logic to work with double precision.

+++ MatrixMulMP/MatrixMulMP.c
@@ -5,9 +5,9 @@
 
 FILE *pFile;		//File handler for matrix files
 
-float* matrixA = NULL;	//Pointer to matrix A
-float* matrixB = NULL;	//Pointer to matrix B
-float* matrixC = NULL;	//Pointer to matrix C
+double* matrixA = NULL;	//Pointer to matrix A
+double* matrixB = NULL;	//Pointer to matrix B
+double* matrixC = NULL;	//Pointer to matrix C
 
 //Variables to save exec time
 time_t start, end;
@@ -18,7 +18,7 @@ char fName[100];	//Name of the file
 int rA, cA, rB, cB, rC, cC = 0;
 
 //Print matrix
-void printMatrix(float* matrix, int nrow, int ncol) {
+void printMatrix(double* matrix, int nrow, int ncol) {
 	for (int i = 0; i < nrow; i++) {
 		for (int j = 0; j < ncol; j++)
 			printf("% 4.2f ", *(matrix + i * ncol + j));
@@ -27,13 +27,13 @@ void printMatrix(float* matrix, int nrow, int ncol) {
 }
 
 // Fill matrix from console
-void fillMatrixUI(char* fName, float* matrix, int rows, int cols){
+void fillMatrixUI(char* fName, double* matrix, int rows, int cols){
 	err = fopen_s(&pFile, fName, "r");
 	if (err == 0){
 		for (int i = 0; i < rows; i++)
 			for (int j = 0; j < cols; j++)
 				if (!feof(pFile)) {
-					fscanf_s(pFile, "%f", (matrix + i * cols + j));
+					fscanf_s(pFile, "%lf", (matrix + i * cols + j));
 				}
 				else {
 					printf("Elementos insuficientes en el archivo, saliendo...\n");
@@ -60,7 +60,7 @@ int main(void){
 		printf("Introduzca las dimensiones de la matriz A (filas columnas): ");
 		scanf_s("%d %d", &rA, &cA);
 		strcpy_s(fName, 100, "D:\\Users\\Alessandro Balcazar\\Documents\\Proyecto_Multi\\MatrixMulMP\\Debug\\matrixA2500.txt");
-		matrixA = (float*)malloc(sizeof(float) * rA * cA);
+		matrixA = (double*)malloc(sizeof(double) * rA * cA);
 		fillMatrixUI(fName, matrixA, rA, cA);
 
 		/*-----------------------------Llenado de matriz B-----------------------------*/
@@ -68,7 +68,7 @@ int main(void){
 		scanf_s("%d %d", &rB, &cB);
 		strcpy_s(fName, 100, "D:\\Users\\Alessandro Balcazar\\Documents\\Proyecto_Multi\\MatrixMulMP\\Debug\\matrixB2500.txt");
 		err = fopen_s(&pFile, fName, "r");
-		matrixB = (float*)malloc(sizeof(float) * rB * cB);
+		matrixB = (double*)malloc(sizeof(double) * rB * cB);
 		fillMatrixUI(fName, matrixB, rB, cB);
 
 		/*-----------Validacion para poder multuplicar las matrices-----------*/
